# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sizes": {
      "post": {
        "operationId": "SizesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSizeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Sizes"
        ]
      },
      "get": {
        "operationId": "SizesController_findAll",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Sizes"
        ]
      }
    },
    "/sizes/{id}": {
      "get": {
        "operationId": "SizesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Sizes"
        ]
      },
      "patch": {
        "operationId": "SizesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSizeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Sizes"
        ]
      },
      "delete": {
        "operationId": "SizesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Sizes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ecom API",
    "description": "E-commerce REST API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSizeDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSizeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}